requirements/tests.txt
tox>=1.7.0
pytest>=5.4.3
pytest-django>=3.9.0
pytest-cov>=2.10.0
